+++ lib/src/lib.rs
@@ -69,6 +69,7 @@ pub mod repo;
 pub mod repo_path;
 pub mod revset;
 pub mod revset_graph;
+mod revset_parser;
 pub mod rewrite;
 pub mod settings;
 pub mod signing;
